% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{abs_path}
\alias{abs_path}
\title{Make absolute path from file/dir path parts}
\usage{
abs_path(...)
}
\arguments{
\item{...}{Character vectors of path parts.}
}
\value{
Character string
}
\description{
Make absolute path from file/dir path parts
}
\examples{
abs_path(".")
abs_path("..", "..")
}
\seealso{
\code{\link{file.path}}, \code{\link{normalizePath}}
}
