% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general.R
\name{age_months}
\alias{age_months}
\title{Age calculater (months)}
\usage{
age_months(birthDate, refDate = Sys.Date())
}
\arguments{
\item{birthDate}{Date of birth string that the function \code{\link{ymd}} can understand. Typically in the format "yyyy-mm-dd" or "yyyy/mm/dd"}

\item{refDate}{Reference date string. Either today's date or some other time after DOB. Defaults to today.}
}
\value{
Numeric value of age in months
}
\description{
Age calculater (months)
}
\examples{
age_months("2013-01-10")
}

