% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{ageCalculator}
\alias{ageCalculator}
\title{Age calculater (months)}
\usage{
ageCalculator(dob, ref, lub.fmt = lubridate::mdy)
}
\arguments{
\item{dob}{Date of birth string that the function \code{\link{ymd}} and 
others like it can understand. Typically in the format "yyyy-mm-dd" or "yyyy/mm/dd"}

\item{ref}{Reference date string. Either today's date or some other time after 
\code{dob} Defaults to today.}

\item{lub.fmt}{Lubridate function for the input dates, such as 
\code{\link{ymd}} or any function that returns a \code{POSIXct} format. 
Defaults to \code{\link{mdy}}}
}
\value{
Numeric value of age in months
}
\description{
Calculates ages in months with decimal days from date of birth until up to some point
}
\details{
If you're going to use a reference date, make sure the format for both dob 
and ref are the same. For example, don't use ymd for dob and mdy for ref. 
You'll get wrong values.
}
\examples{
ageCalculator("01-10-2013")
ageCalculator(c("05-13-1983", "01-10-2013"), c("05-13-2000", "10-07-2014"))
ageCalculator("2013/01/10", lub.fmt=lubridate::ymd)
}

