% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{annotate_corner}
\alias{annotate_corner}
\title{Write text to one of four corners of a plot}
\usage{
annotate_corner(text, pos = "tr", ...)
}
\arguments{
\item{text}{character string}

\item{pos}{character of \code{'tl'}, \code{'tr'}, \code{'bl'}, \code{'br'} to indicate position}

\item{...}{additional options passed to \code{ggplot2::annotate}}
}
\value{
ggplot layer
}
\description{
Write text to one of four corners of a plot
}
\examples{
example_plot(facets=FALSE)+annotate_corner('Hi.')
}
