% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{append_args}
\alias{append_args}
\title{Update a list of arguments with new values, with possible overwrite}
\usage{
append_args(arglist, updates = NULL, keep_original = TRUE)
}
\arguments{
\item{arglist}{list of input arguments}

\item{updates}{list of updates to append}

\item{keep_original}{If set to \code{TRUE} (default), items already existing in \code{arglist}
will not be changed.}
}
\value{
list
}
\description{
Update a list of arguments with new values, with possible overwrite
}
\examples{
arglist <- list(x = 1, y = 2, z = 3)
append_args(arglist, list(z = NA, w = 0))
append_args(arglist, list(z = NA, w = 0), FALSE)
}
