% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{argval2char}
\alias{argval2char}
\title{Convert input arg values to character vector}
\usage{
argval2char(...)
}
\arguments{
\item{...}{input args}
}
\value{
character vector
}
\description{
Convert input arg values to character vector
}
\examples{
argval2char(x='a string', y=2*pi, z=NA, f = ~ x + b)
}
