% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{axis_limit}
\alias{axis_limit}
\title{Make axis limits from a range of values}
\usage{
axis_limit(xrange, d = 2, e = 0)
}
\arguments{
\item{xrange}{A range of values. Can be found with \link{range}}

\item{d}{Number of digits to round to.}

\item{e}{Expansion multiplier. Suggests something like 0.1, or 0.05. Defaults to 0.}
}
\description{
This is normally used for plotting, where it expands the range of values and also truncates digits.
}
\examples{
axis_limit(xrange=c(100.1234,200.4321), d=1, e=0.1)
}
\seealso{
\link{range}

Other graphics: \code{\link{alpha_override}},
  \code{\link{getHCL}}, \code{\link{getLuminance}},
  \code{\link{margin_text}}, \code{\link{rainbow_colors}},
  \code{\link{theme_mejr}}
}
