% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{beta_moments}
\alias{beta_moments}
\title{Beta distribution moments}
\usage{
beta_moments(a, b, mu, sigma)
}
\arguments{
\item{a}{alpha parameter. Must be greater than zero.}

\item{b}{beta parameter. Must be greater than zero.}

\item{mu}{mean parameter}

\item{sigma}{standard deviation parameter. Must be less than (mu*(1-mu))^2}
}
\value{
a list of stats
}
\description{
Beta distribution moments
}
\examples{
beta_moments(2, 0.5)
beta_moments(mu = 0.7, sigma = 0.1)
}
