% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{chol2cov}
\alias{chol2cov}
\title{Cholesky correlation matrix and standard deviations to covariance matrix}
\usage{
chol2cov(sigma, cm, tcross = FALSE)
}
\arguments{
\item{sigma}{sigma vector}

\item{cm}{cholesky factor correlation matrix}

\item{tcross}{set to TRUE to use tcrossprod instead of crossprod. That is, if cm is flipped as in STAN.}
}
\value{
matrix, covariances
}
\description{
Cholesky correlation matrix and standard deviations to covariance matrix
}
\examples{
# correlation matrix
m <- matrix(c(1,-0.15,0.67,-0.15,1,.2,0.67,.2,1), ncol=3)

# standard deviations on the diagonal
sigma <- c(.5, 1, 1.5)

# R formatted cholesky factor
cm <- chol(m)
chol2cov(sigma, cm)

# STAN formatted cholesky factor
cm <- stan_chol(m)
chol2cov(sigma, cm, TRUE)
}

