% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{class_override}
\alias{class_override}
\title{Override column classes}
\usage{
class_override(x, class_list)
}
\arguments{
\item{x}{a data.frame or data.table}

\item{class_list}{a list of class names with each list item containing column names to assign classes to}
}
\value{
same x but with new classes. If data.table, does not copy.
}
\description{
Override column classes
}
\examples{
class_list <- list(character = c("V1", "V2"), double = "V3")
x <- data.frame(V1 = letters[3:1], V2 = 1:3, V3 = 1:3)

# view current classes, note V1 is a factor
sapply(x, class)

x$V1 <- as.character(x$V1)

# change classes
x <- class_override(x, class_list)

# view again
sapply(x, class)
}

