% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{clear_ws}
\alias{clear_ws}
\title{Clear workspace}
\usage{
clear_ws(hidden = TRUE, env = .GlobalEnv)
}
\arguments{
\item{hidden}{Removes hidden objects. Logical value. DEFAULT=\code{TRUE}.}

\item{env}{Specify environment which to remove objects from.
DEFAULT=\code{.GlobalEnv}.}
}
\value{
NA
}
\description{
This will clear all objects found in the Global Environment by default. It
also clears hidden objects (anything with a \code{.})
}
\details{
If you wanted to clear anything else besides Global, specify the environment
with the \code{env} argument
}
\examples{
clear_ws()
}
\seealso{
Other helpers: \code{\link{fac2num}},
  \code{\link{has_data}}, \code{\link{print_sec}},
  \code{\link{unload_pkg}}
}
\concept{helpers}
