% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{color_override}
\alias{color_override}
\title{Change default colors}
\usage{
color_override(n = 4, values = NULL, fill = FALSE)
}
\arguments{
\item{n}{number of colors to choose}

\item{values}{changes colors using values in values}

\item{fill}{changes 'fill' property. Otherwise changes 'color' property.}
}
\value{
ggproto obj
}
\description{
Change default colors
}
\examples{
ex_plot()+color_override(8)
}
