% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{combine_plots}
\alias{combine_plots}
\title{Combine multiple ggplots into one plot}
\usage{
combine_plots(..., plots, layout, heights, widths, ncols, show = TRUE)
}
\arguments{
\item{...}{names of plot objects}

\item{plots}{a list of plot objects if not using ...}

\item{layout}{custom plot layout}

\item{heights}{ratio of heights per row}

\item{widths}{ratio of widths per column}

\item{ncols}{optionally specify number of columns instead of layout}

\item{show}{Print plot or just return gtable object}
}
\value{
gtable
}
\description{
Combine multiple ggplots into one plot
}
\examples{
my_plots <- lapply(1:5, function(i) ggplot2::qplot(rnorm(100)))
combine_plots(plots = my_plots)
combine_plots(plots = my_plots,
 layout = matrix(c(1:5,5), ncol=2, byrow = TRUE),
 heights = c(.4,.4,.2),
 widths = c(.6,.4))
}
