% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{dots2list}
\alias{dots2list}
\title{Convert dots to list}
\usage{
dots2list(...)
}
\arguments{
\item{...}{a set of inputs to convert}
}
\description{
Convert dots to list
}
\examples{
dots2list(x='a string', y=2*pi, z=NA, f = ~ x + b)
}
\seealso{
symbol2char
}
