% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{dprime}
\alias{dprime}
\title{Calculate d'}
\usage{
dprime(h, f)
}
\arguments{
\item{h}{Hit rates}

\item{f}{False alarm rates}
}
\description{
Calculate d' from hit and false alarm rates
}
\examples{
dprime(.47, .01)
dprime(.51, 0)
}
