% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafunctions.R
\name{dsplit}
\alias{dsplit}
\title{data split into lists}
\usage{
dsplit(data, ...)
}
\arguments{
\item{data}{a data.frame or data.table}

\item{...}{unquoted column names}
}
\value{
a list of data.table objects
}
\description{
data split into lists
}
\examples{
data <- cars
dsplit(data, speed)
}
