% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{empty_str}
\alias{empty_str}
\title{Check if a character string/vec is empty}
\usage{
empty_str(x)
}
\arguments{
\item{x}{character string or vector of strings}
}
\value{
logical value
}
\description{
Checks include NULL, "", or character vectors of length 0.
}
\examples{
# all empty
empty_str(NULL)
empty_str("")
empty_str(character())
empty_str(c("", ""))

# not empty, one string contains a space
empty_str(c("", " "))
}
