% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ex_plot}
\alias{ex_plot}
\title{Make a plot example from mejr theme}
\usage{
ex_plot(facets = TRUE, ax = "bottom", ay = "left", switch = NULL)
}
\arguments{
\item{facets}{use plotting facets}

\item{ax}{x axis position}

\item{ay}{y axis position}

\item{switch}{switch facet panels and axes}
}
\description{
Uses the ggplot2 diamonds data set as an example
}
\examples{
ggplot2::theme_set(ggdistribute::theme_mejr(debug_text = TRUE))
ex_plot()
}
