% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan.R
\name{extract_partial}
\alias{extract_partial}
\title{Extract but without some of the chains}
\usage{
extract_partial(object, pars, chains, print = FALSE)
}
\arguments{
\item{object}{Stan object that was previously saved}

\item{pars}{Parameter names, as a vector of character values}

\item{chains}{Chain numbers to keep, as a vector of integers}

\item{print}{Print new stats using the kept chains}
}
\value{
list with each sublist as a chain, plus a new summary
}
\description{
Extract but without some of the chains
}
\examples{
stanfit <- rstan_mejr()
prams <- extract_partial(stanfit$stan_mcmc, "Beta", c(1,3), TRUE)
}

