\name{f2num}
\alias{f2num}
\title{Factor to numeric}
\usage{
f2num(x)
}
\arguments{
  \item{x}{Factor column}
}
\value{
Numeric vector
}
\description{
Takes a factor and tries to coerce to numeric. Helpful if
numbers have been coverted to factors.
}
\seealso{
Other helpers: \code{\link{cp}}; \code{\link{has.data}};
\code{\link{loadpkg}}
}

