% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general.R
\name{fac2num}
\alias{fac2num}
\title{Factor to numeric class}
\usage{
fac2num(x)
}
\arguments{
\item{x}{Factor column}
}
\value{
Numeric vector
}
\description{
Takes a factor and tries to coerce to numeric. Helpful if numbers have been coverted to factors.
}
\seealso{
Other helpers: \code{\link{clrAll}}; \code{\link{hasData}};
  \code{\link{loadPkg}}; \code{\link{printSec}};
  \code{\link{rmPkg}}
}

