% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{format_vector}
\alias{format_vector}
\title{Pretty formatting for a numeric vector}
\usage{
format_vector(x, nrows = NULL, round = getOption("mejr.round"))
}
\arguments{
\item{x}{vector of integers or floating point values}

\item{nrows}{split vector into this many rows instead of a single row vector}

\item{round}{round to n significant digits}
}
\value{
a formatted numeric vector as string
}
\description{
Pretty formatting for a numeric vector
}
\examples{
format_vector(rnorm(10), 2, 3)
}
