% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{gamma_stats}
\alias{gamma_stats}
\title{gamma distribution stats}
\usage{
gamma_stats(shape, rate)
}
\arguments{
\item{shape}{shape parameter of gamma distribution}

\item{rate}{rate parameter of gamma distribution}
}
\value{
a vector of statistics
}
\description{
gamma distribution stats
}
\examples{
gamma_stats(1, 2)
}

