% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{generate_doc}
\alias{generate_doc}
\title{Render an Rmarkdown document}
\usage{
generate_doc(filename, format = "html_document", ...)
}
\arguments{
\item{filename}{character input specifying markdown file name}

\item{format}{character name of format function from the \code{rmarkdown} package,
e.g., \code{c('html_document', 'html_document')}}

\item{...}{additional options passed to rmarkdown::render}
}
\description{
Wrapper for formatting options from rmarkdown::*_document() functions
}
\examples{
generate_doc('input.Rmd', 'pdf_document')
}
