% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_colors}
\alias{get_colors}
\title{Get colors from Brewer pallette}
\usage{
get_colors(n = 11)
}
\arguments{
\item{n}{number of colors to generate}
}
\value{
A character vector of hex color codes
}
\description{
Get colors from Brewer pallette
}
\examples{
get_colors(16)
}

