% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{getcrf}
\alias{getcrf}
\title{Get current R file path}
\usage{
getcrf(parent = TRUE, pos = 1L)
}
\arguments{
\item{parent}{Return R file parent directory. Logical.}
}
\value{
character
}
\description{
Get current R file path
}
\examples{
# to return directory of file the function is being called from.
getcrf()

# to return file path
getcrf(FALSE)
}
