% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{\%?n\%}
\alias{\%?n\%}
\title{Check if an object on the LHS has ALL the names on the RHS character vector}
\usage{
obj \%?n\% names
}
\arguments{
\item{obj}{data object with a names method}

\item{names}{character vector of names}
}
\description{
Check if an object on the LHS has ALL the names on the RHS character vector
}
\examples{
d <- data.frame(x=1, y=2)
d \%?n\% 'x'              # <- TRUE
d \%?n\% c('x', 'y')      # <- TRUE
d \%?n\% 'z'              # <- FALSE
d \%?n\% c('x', 'z')      # <- FALSE
d \%?n\% c('x', 'y', 'z') # <- FALSE
}
