% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafunctions.R
\name{has_data}
\alias{has_data}
\title{Check for empty data frames or vectors}
\usage{
has_data(obj)
}
\arguments{
\item{obj}{The object to be evaluated}
}
\value{
Logical
}
\description{
This will check to see if a data frame, vector, or matrix has data in it (not
empty). If so, returns TRUE
}
\examples{
x <- character()
has_data(x)

x <- data.frame(V1=character())
has_data(x)
}
\seealso{
Other helpers: \code{\link{clear_ws}},
  \code{\link{fac2num}}, \code{\link{print_sec}},
  \code{\link{unload_pkg}}
}
\keyword{empty}
