% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{heat_colors}
\alias{heat_colors}
\title{A set of colors from Kindlmann space}
\usage{
heat_colors(n, bias = 1)
}
\arguments{
\item{n}{an integer of the number of colors to get}

\item{bias}{a positive number. Higher values give more widely spaced colors at the high end.}
}
\value{
Hex Codes
}
\description{
A set of colors from Kindlmann space
}
\examples{
heat_colors(10)
}

