% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{interleave}
\alias{interleave}
\title{interleave vectors or matrices by alternating sequential values}
\usage{
interleave(..., dim = 2)
}
\description{
interleave vectors or matrices by alternating sequential values
}
