% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{kwargs_keys}
\alias{kwargs_keys}
\title{Get the named argument names from dots}
\usage{
kwargs_keys(..., named_only = TRUE)
}
\arguments{
\item{...}{args passed to function call}
}
\value{
character vector
}
\description{
Get the named argument names from dots
}
\examples{
kwargs_keys(x = 1, y = NULL, F, 0.0)
kwargs_keys(x = 1, y = NULL, F, 0.0, named_only = FALSE)
}
