% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{label_plot}
\alias{label_plot}
\title{Add labels to existing plot}
\usage{
label_plot(labels, x, y, g = list(fontsize = 14, fontface = "bold"), ...)
}
\arguments{
\item{labels}{character vector of labels to use}

\item{x}{horz positions of items in labels}

\item{y}{vert positions of items in labels}

\item{g}{list of options passed to \code{grid::gpar}}

\item{...}{optional args passed to \code{grid::grid.text}}
}
\value{
NULL. prints to current graphics device.
}
\description{
Add labels to existing plot
}
\examples{
example_plot()
label_plot(c('a label', 'another one'), c(.2, .8), c(.333, .667))

# use extra options from grid::grid.text
label_plot('last one', 0.5, 0.5, just='center')
}
