% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafunctions.R
\name{list2excel}
\alias{list2excel}
\title{Write a list of data frames to an Excel file}
\usage{
list2excel(excel_list, filename, n_chunk_cols = Inf)
}
\arguments{
\item{excel_list}{a data.frame/table or list of them}

\item{filename}{filename to save workbook to}

\item{n_chunk_cols}{if multiple data sets on one sheet, specify number of
cols for each subset}
}
\description{
Write a list of data frames to an Excel file
}
\examples{
excel_list <- list(iris=iris, apply(iris3, 3, as.data.frame))
list2excel(excel_list, '~/test_mejr_excel.xlsx', n_chunk_cols=2)
}
