% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{list_files}
\alias{list_files}
\title{list.files wrapper}
\usage{
list_files(x = ".", ext = ".*", recursive = TRUE)
}
\arguments{
\item{x}{directory string}

\item{ext}{file extension to search for}

\item{recursive}{include subdirectories}
}
\value{
character vector
}
\description{
List all files in a directory recursively using full paths
}
\examples{
list_files()
list_files(ext='.R')
}
