% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{logaddexp}
\alias{logaddexp}
\title{Log add exponentials}
\usage{
logaddexp(x1, x2)
}
\arguments{
\item{x1}{scalar or input vector}

\item{x2}{scalar or input vector}
}
\value{
scalar or input vector
}
\description{
Log add exponentials
}
\examples{
logaddexp(1, 1) # log(2.718282 + 2.718282)
logaddexp(0, c(.01, .5, .9, 2))
}
