% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{logistic}
\alias{logistic}
\title{Logistic function}
\usage{
logistic(x, l = 1, k = 2, m = 0)
}
\arguments{
\item{x}{logit scaled input vector}

\item{l}{max possible value}

\item{k}{slope. Default=2. Set to 1 to be equal to sigmoid function.}

\item{m}{mean or center point of x}
}
\value{
numeric vector or value
}
\description{
Logistic function
}
\examples{
x <- logit(seq(0.001, 0.999, length.out = 50))
plot(x = x, y = logistic(x), type = "l")             # defaults
plot(x = x, y = logistic(x, k=1), type = "l")        # sigmoid
plot(x = x, y = logistic(x, l=2, k=8)-1, type = "l") # change range and slope then subtract 1
}
