% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{logit}
\alias{logit}
\title{Logit function}
\usage{
logit(p)
}
\arguments{
\item{p}{Numeric value between 0 and 1, a probability}
}
\value{
Numeric values ranging from -Inf to Inf
}
\description{
The logit, or inverse of the sigmoid (logistic) function.
}
\examples{
logit(0.5)
logit(.01)
logit(1)

p <- seq(1e-04, 0.9999, 1e-04)
plot(x = p, y = logit(p), type = "l")
}
