% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-misc.R
\name{luminance}
\alias{luminance}
\title{Get perceptual luminance estimate from RGB values}
\usage{
luminance(rgb)
}
\arguments{
\item{rgb}{numeric vector of RGB values. Example: \code{c(0,127,255)}}
}
\description{
This will calculate an estimate of percieved luminance when provided a
vector of RGB values, in that order.
}
\examples{
getLuminance(c(0,127,255))
getLuminance(as.vector(col2rgb("gray80")))
}
\seealso{
\link{rgb} \link{col2rgb} \link{hcl}

Other graphics: \code{\link{alpha_override}},
  \code{\link{theme_mejr}}
}
