% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datafunctions.R
\name{makeEmptyDf}
\alias{makeEmptyDf}
\title{Creates an empty dataframe from column names}
\usage{
makeEmptyDf(cnames)
}
\arguments{
\item{cnames}{Character vector of column names. If none provided, 5 columns V1-V5 will be used instead.}
}
\description{
If provided a list of character names, it will create a dataframe with no rows but the names of your columns
Each column defaults to class of type \code{numeric}.
}
\examples{
makeEmptyDf(c("Subject","score"))
}

