% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{margin_text}
\alias{margin_text}
\title{Draw text on the left or right margin of a plot}
\usage{
margin_text(gplot, text, y, side = "right", margin = 1, cex = 0.75, ...)
}
\arguments{
\item{gplot}{the ggplot object}

\item{text}{a character vector of text to write for each value of y}

\item{y}{the coordinates along the y-axis which to write the text}

\item{side}{choose either "left" or "right" side of the plot}

\item{cex}{text rescale}

\item{...}{additional graphical parameters}
}
\description{
This will increase the size of the margin then draw text that you specify at some points along the y-axis.
}
\examples{
margin_text(p1, text=c("text1", "text2"), y=c(-1, 1))
}
\seealso{
Other graphics: \code{\link{alpha_override}},
  \code{\link{axis_limit}}, \code{\link{getHCL}},
  \code{\link{getLuminance}}, \code{\link{rainbow_colors}},
  \code{\link{theme_mejr}}
}
