% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{math_matrix}
\alias{math_matrix}
\title{Print a LaTeX math matrix from an R matrix object}
\usage{
math_matrix(sym = "", x, rows = NULL, cols = NULL,
  round = getOption("mejr.round"), print = TRUE)
}
\arguments{
\item{sym}{escaped LaTeX math expression to print before the matrix}

\item{x}{matrix object}

\item{rows}{n rows to print (defaults to all)}

\item{cols}{n cols to print (defaults to all)}

\item{round}{n significant digits to print (defaults to 2)}

\item{print}{immediately print or return text object}
}
\value{
character
}
\description{
Print a LaTeX math matrix from an R matrix object
}
\examples{
x <- matrix(rnorm(25), 5)
math_matrix('\\\\bar{x}=', x, 4, 4, 2)
}
