% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mejr.R
\docType{package}
\name{mejr}
\alias{mejr}
\alias{mejr-package}
\title{mejr: Package}
\description{
This package has no main purpose but to make my life easier
instead of copying and pasting functions from different folders.
}
\section{Info}{

See \code{library(help="mejr")} for a list of functions. \cr
\cr
Below are some useful links I can click on instead of google searching them: \cr
\cr
\url{http://docs.ggplot2.org/current/}
}

\section{Function list}{

\code{\link{age_months}} \cr
\code{\link{alpha_override}} \cr
\code{\link{axisLim}} \cr
\code{\link{categorize}} \cr
\code{\link{clrAll}} \cr
\code{\link{denseMode}} \cr
\code{\link{dprime}} \cr
\code{\link{examplePlot}} \cr
\code{\link{fac2num}} \cr
\code{\link{getHCL}} \cr
\code{\link{getLuminance}} \cr
\code{\link{hasData}} \cr
\code{\link{hdi}} \cr
\code{\link{hdiq}} \cr
\code{\link{loadPkg}} \cr
\code{\link{logit}} \cr
\code{\link{logmean}} \cr
\code{\link{mPlot}} \cr
\code{\link{makeEmptyDf}} \cr
\code{\link{marginText}} \cr
\code{\link{mejrColor}} \cr
\code{\link{ms2frames}} \cr
\code{\link{normalize}} \cr
\code{\link{plotPDF}} \cr
\code{\link{pram_hist}} \cr
\code{\link{printSec}} \cr
\code{\link{rainbow2}} \cr
\code{\link{rmPkg}} \cr
\code{\link{rstan_mejr}} \cr
\code{\link{sigmoid}} \cr
\code{\link{snap_range}} \cr
\code{\link{stackCSV}} \cr
\code{\link{stan_chol}} \cr
\code{\link{stan_fit_stat}} \cr
\code{\link{stddev_ME}} \cr
\code{\link{theme_mejr}} \cr
\code{\link{trim}} \cr
\code{\link{varcov_ME}} \cr
\code{\link{view_stan_chains}} \cr
\code{\link{zMat}}
}

