% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{mejrColor}
\alias{mejrColor}
\title{A set of colors from Kindlmann space}
\usage{
mejrColor(n, bias = 1)
}
\arguments{
\item{n}{an integer of the number of colors to get}

\item{bias}{a positive number. Higher values give more widely spaced colors at the high end.}
}
\value{
Hex Codes
}
\description{
A set of colors from Kindlmann space
}
\examples{
mejrColor(10)
}

