% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mPlot}
\alias{mPlot}
\title{print multiple ggplots in a single graphics device}
\usage{
mPlot(..., plotlist, cols, layout, h, w)
}
\arguments{
\item{...}{Each plot to be merged one after another}

\item{plotList}{Alternative list of plots instead of using the ...}

\item{cols}{Number of columns for final plot}

\item{layout}{matrix indicating layout of plot structure}

\item{h}{vector of heights}

\item{w}{vector of widths}
}
\description{
This will take a vector of ggplot2 objects and then lay them out, one-by-one, in a single plot.
}
\details{
If \code{cols} is specified, it will wrap plots along the number of columns.
Alternatively, you can use a custom \link{layout}.
}
\examples{
\dontrun{

# Make data and plots
dat <- data.frame(y=rnorm(100), x=seq(-2,2, length.out=100))
p1 <- ggplot(dat, aes(x=x,y=y))+geom_point()
p2 <- ggplot(dat, aes(x=x,y=y))+geom_line()

mPlot(p1, p2, cols=2)
mPlot(p1, p2, cols=1, h=c(.25, .75))
mPlot(p1, p2, layout=rbind(1,2), h=c(.25, .75), w=1)

}
}
\seealso{
layout

Other graphics: \code{\link{alpha_override}};
  \code{\link{axisLim}}; \code{\link{getHCL}};
  \code{\link{getLuminance}}; \code{\link{mejrColor}};
  \code{\link{theme_mejr}}
}
\keyword{ggplot2}
\keyword{theme_set}

