% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ms2frames}
\alias{ms2frames}
\title{Convert milliseconds to frame number}
\usage{
ms2frames(x, fps = 30, tstart = 0, mscut = FALSE)
}
\arguments{
\item{x}{millisecond numeric or integer data}

\item{fps}{frames per second. Single value.}

\item{tstart}{time to be used as the initial level in a factor. Assumes 0 time.}

\item{mscut}{return the cutoff points in milliseconds instead of frame number.
Similar to the millisecond vector that was entered but now binned to a specific value.}
}
\description{
Provide the frame rate to use to create break points from millisecond time data
}
\details{
Assumes time starts at 0, but this can be changed.
}
\examples{
# sequence of milliseconds
x <- seq(1, 1009, 12)

# 30 fps video
ms2frames(x, fps=30)

# first frames are zero until start frame is encountered
ms2frames(x, fps=29.97, tstart=333)


names(x) <- sprintf("\%.2f", ms2frames(x, fps=30, mscut=TRUE))
}

