% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datafunctions.R
\name{multi_merge}
\alias{multi_merge}
\title{Multiple data merge}
\usage{
multi_merge(data_list, setkeys = FALSE, ...)
}
\arguments{
\item{data_list}{List of separate data.frames/tables to merge}

\item{...}{Additional arguments passed to \code{merge}}

\item{setkey}{use \code{setkey} on all columns in a data.table before merging.}
}
\value{
A data.frame/data.table, depending on the input type in the list
}
\description{
Merge multiple data sets from a list
}
\details{
All items in the list must be the same kind, either all data.table or all data.frame.
}
\examples{
d1 <- data.table(1:4, letters[1:4])
d2 <- data.table(1:3, letters[3:5])
d3 <- data.table(1:5, letters[1:5])
data_list <- list(d1, d2, d3)
merged_data <- multi_merge(data_list, by=c("V1", "V2"), all = TRUE)
merged_data <- multi_merge(data_list, setkeys = TRUE, all = TRUE)
}

