% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{na_slices}
\alias{na_slices}
\title{Adds rows or columns of NAs to a matrix}
\usage{
na_slices(x, head = NULL, d = 1)
}
\arguments{
\item{x}{matrix object}

\item{head}{number of rows or columns to keep, or index of which to keep}

\item{d}{dimension. 1 = rows, 2 = columns}
}
\value{
matrix
}
\description{
Adds rows or columns of NAs to a matrix
}
\examples{
x <- diag(5)
na_slices(x, 4)
na_slices(x, 4, 2)
}
