% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{new_rproject}
\alias{new_rproject}
\title{Make a new project directory tree}
\usage{
new_rproject(name, root_dir = ".")
}
\arguments{
\item{name}{New project's name}

\item{root_dir}{Destination folder where to create the project directory}
}
\description{
Make a new project directory tree
}
\examples{
new_rproject("MyRProject", "~")
}
