% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{non_null}
\alias{non_null}
\title{check if all values are NULL or possibly all NAs, return a single value if so.}
\usage{
non_null(x, val, check_na = FALSE)
}
\description{
check if all values are NULL or possibly all NAs, return a single value if so.
}
