% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{normalize}
\alias{normalize}
\title{Normalize a vector of values}
\usage{
normalize(x, type = c("weights", "relative"))
}
\arguments{
\item{x}{numeric or integer value or vector of values of these types}

\item{type}{sum to one or restrict range from 0 to 1}
}
\description{
Normalize a vector of values
}
\details{
Will normalize to sum to one or adjust values to be between 0 and 1.
}
\examples{
x <- sort(runif(10, -100, 100))
normalize(x, "weights")
normalize(x, "relative")
}

