% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{os_name}
\alias{os_name}
\title{Get OS abbreviation name}
\usage{
os_name()
}
\value{
\code{character}, one of \code{WIN} for Windows, \code{MAC} for Darwin, \code{LNX} for Linux
}
\description{
Get OS abbreviation name
}
\examples{
os_name()
}
