% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{pairwise}
\alias{pairwise}
\title{all pairwise combinations}
\usage{
pairwise(n)
}
\arguments{
\item{n}{set size (integer)}
}
\value{
n pairs by 2 matrix
}
\description{
all pairwise combinations
}
\examples{
pairwise(3)
}
