% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_spread.R
\name{position_spread}
\alias{position_spread}
\title{PositionDodge but vertically}
\usage{
position_spread(height = NULL, preserve = c("total", "single"))
}
\arguments{
\item{height}{similar to \code{width} from position_dodge}

\item{preserve}{see position_dodge}
}
\description{
PositionDodge but vertically
}
\examples{
geom_point(aes(group=myGroup), position_spread(height = 0.5))
}
