% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pram_hist}
\alias{pram_hist}
\title{Point estimates and histogram plots of fitted parameters in Stan}
\usage{
pram_hist(x, bndw = 0.33, fname = "plot_stanfit_hist.pdf")
}
\arguments{
\item{x}{rstan object}

\item{bndw}{adjust density line bandwidth}

\item{fname}{pdf file name for histograms}
}
\description{
Plots histograms, densities, and central tendency (defaults to median)
}
\examples{
-
}

