% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{print_fmt}
\alias{print_fmt}
\title{sprintf format string}
\usage{
print_fmt(round = getOption("mejr.round"),
  digits = getOption("mejr.digits"), format = "g", leading = TRUE)
}
\description{
sprintf format string
}
