% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{rainbow_colors}
\alias{rainbow_colors}
\title{Return a set of custom rainbow themed colors}
\usage{
rainbow_colors(n = 1, adj = 0, reverse = FALSE, fullrange = FALSE,
  alpha = 1)
}
\arguments{
\item{n}{Number of distinct colors to retreive. If more than one, they will be equally spaced.}

\item{adj}{Rotate the wheel by a certain amount [0,360]}

\item{reverse}{Whether to reverse the color order or not}

\item{fullrange}{Whether to use the full wheel or not. Defaults to [80,300]}

\item{alpha}{Transparency value from 0 to 1. Defaults to 1 (opaque).}
}
\description{
This function is similar to \link{rainbow}, but with my own defaults
}
\examples{
\dontrun{
rainbow_colors(10)
}
}
\seealso{
\link{rainbow}

Other graphics: \code{\link{alpha_override}},
  \code{\link{axis_limit}}, \code{\link{getHCL}},
  \code{\link{getLuminance}}, \code{\link{margin_text}},
  \code{\link{theme_mejr}}
}
