% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rcov}
\alias{rcov}
\title{random covariance matrix}
\usage{
rcov(k, sigma, eta, sign)
}
\arguments{
\item{k}{dimensions of matrix}

\item{sigma}{standard deviations of diagonal elements}

\item{eta}{correlation adjuster > 0}

\item{sign}{correlation sign}
}
\value{
matrix
}
\description{
random covariance matrix
}
\examples{
rcov(3)
}

