% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{reformatSelectionAddin}
\alias{reformatSelectionAddin}
\title{Reformat a block of code using formatR. Map to keyboard shortcut.}
\usage{
reformatSelectionAddin()
}
\description{
Reformat a block of code using formatR. Map to keyboard shortcut.
}
