% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mejr-internal.R
\name{require_pkg}
\alias{require_pkg}
\title{Check if package is installed quietly. Load namespace.}
\usage{
require_pkg(str)
}
\arguments{
\item{str}{package name}
}
\description{
Check if package is installed quietly. Load namespace.
}
\examples{
require_pkg('ggplot2')
}
