% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{reverse_scale}
\alias{reverse_scale}
\title{Reverse of scale}
\usage{
reverse_scale(x, m, s)
}
\arguments{
\item{x}{scaled numeric vector}

\item{m}{mean of original scale}

\item{s}{standard deviation of original scale}
}
\value{
numeric vector
}
\description{
Takes mean and standard devation then gets unstandardized vector
}
\examples{
x <- rpois(30, 100)
m <- mean(x)
s <- sd(x)
x <- scale(x)
reverse_scale(x, m, s)
}
