% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{rfmtSelectionAddin}
\alias{rfmtSelectionAddin}
\title{Reformat a block of code using rfmt. Map to keyboard shortcut.}
\usage{
rfmtSelectionAddin(win_cygwin = getOption("mejr.cygwin"))
}
\arguments{
\item{win_cygwin}{root cygwin path}
}
\description{
Reformat a block of code using rfmt. Map to keyboard shortcut.
}
\examples{
# Code to test for reformatting
someTestFunc <- function() {
    library(magrittr)
    library(data.table)
    X <- as.data.frame(cars)
    new.data <- X[, names(X), drop = FALSE]
    if (any(new.data$speed > 0 && with(X, identity(dist)) < 5)) {
        Y <- gsub("xxx", paste("yyyy=", new.data$speed, sep = ""), new.data$dist)
    }
    X <- data.table(cars)
    apply(X, 1, function(x) {
        j <- sum(x)
        mean(x)/j
    })
    X[speed > 1, dist, speed] \%>\%
 .[speed > 1, dist, speed] \%>\%
 .[speed > 1, dist, speed] \%>\%
 .[speed > 1, dist, speed] \%>\%
 .[speed > 1, `:=`(x = NA, y = as.character(dist)), speed] \%>\%
 .[speed > 1, `:=`(x = NA), speed]
    ## Make data file
    data.fn <- makeNewDataFile(x = newdata, y = NULL)
    ## Finally, compute forecast for new data
    Z <- .C("forecast", as.character(data.fn), as.character(object$names), as.character(object$data), as.character(object$model), pred = double(nrow(new.data)), output = character(1), PACKAGE = "test")
}
}
