% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan.R
\name{rstan_mejr}
\alias{rstan_mejr}
\title{Stan Mejr}
\usage{
rstan_mejr(model, name, data, pars, samples, init, control, out = NULL,
  parallel = FALSE, debug = FALSE, repackage = NULL, ...)
}
\arguments{
\item{model}{model code or file name, both as character objects}

\item{name}{name of model being fitted}

\item{data}{data in stan format}

\item{pars}{named parameters. Defaults to kept parameters}

\item{samples}{list of sample parameters. See example for list structure.}

\item{init}{initial values}

\item{control}{control paramters for algorithm}

\item{out}{directory for where to save output files}

\item{parallel}{use multiple cores for running multiple chains}

\item{debug}{do a test run with options saved to global env}

\item{repackage}{add additional objects to the output of this function}

\item{...}{other arguments from the main \code{rstan::stan} function}
}
\value{
list of objects: \code{stan_mcmc, pram, central, env, repack}
}
\description{
Wrapper function for simplifying \code{stan}. See \code{?rstan::stan} for more info.
}
\details{
Also provides additional plots and repackages data into a single object
}
\examples{
# Using fake data for example
stan_dat <- rstan_test_data()
stan_fit <- rstan_mejr(data=stan_dat$data,
             model=stan_dat$model,
             samples=list(n_chains = 4, n_final = 1200, n_thin = 2, n_warm = 800),
             pars=stan_dat$pars,
             init=stan_dat$init,
             out="~/Desktop",
             name="example_model",
             parallel=FALSE)

# Run default example model
stan_fit <- rstan_debug <- rstan_mejr(debug=TRUE)
}

