% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan.R
\name{rstan_test_data}
\alias{rstan_test_data}
\title{Stan test data}
\usage{
rstan_test_data(n = 100)
}
\arguments{
\item{n}{number of observations}
}
\value{
list of stan options
}
\description{
Stan test data
}
\examples{
fake_data <- rstan_test_data(n=30)
}

