% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{show_data}
\alias{show_data}
\title{Pretty print a data object or atomic vector}
\usage{
show_data(x, nrows = NULL, caption = NULL, print = TRUE, ...)
}
\arguments{
\item{x}{data object, such as data.frame, data.table, numeric or character vector}

\item{nrows}{number of rows to print}

\item{caption}{table caption text}

\item{print}{print immediately or return string}

\item{...}{other args passed to pander::pandoc.table}
}
\value{
formatted data text
}
\description{
Pretty print a data object or atomic vector
}
\examples{
x <- array(rnorm(100), c(25,5))
colnames(x) <- c("(Intercept)", paste0("V", 1:4))
show_data(x)
show_data(as.data.frame(x))
}
