% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{snapRange}
\alias{snapRange}
\title{Snap a value to either the min or max if outside some range}
\usage{
snapRange(x, l, u)
}
\arguments{
\item{x}{numeric or integer value or vector of values}

\item{l}{lower limit}

\item{u}{upper limit}
}
\description{
If a value lies outside of some range, then this will snap to the limits
}
\details{
Applies to vectors too
}
\examples{
# snaps the vector below to the limits set
x <- c(-2,0,0.5,1, 1.25)
snapRange(x, 0, 1)
}

