% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{softmax}
\alias{softmax}
\title{Softmax}
\usage{
softmax(y, na.rm = TRUE)
}
\arguments{
\item{y}{vector of log values or logit scaled probabilities.}

\item{na.rm}{remove NAs before computing}
}
\value{
numeric vector
}
\description{
Normalize log scaled vector
}
\examples{
softmax(dnorm(seq(-2,2,0.5), log = TRUE))
}
