% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{source_dir}
\alias{source_dir}
\title{Source R files from a directory recursively}
\usage{
source_dir(x = NULL, ...)
}
\arguments{
\item{x}{Character pointing to a valid directory path}

\item{...}{Additional arguments passed to \code{\link{source}}}
}
\description{
Source R files from a directory recursively
}
\examples{
source_dir("path/to/some/folder")
}
\seealso{
\code{\link{source}}
}
