% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan.R
\name{stan_chol}
\alias{stan_chol}
\title{Stan formatted Cholesky factored cov/cor matrix}
\usage{
stan_chol(cmat)
}
\arguments{
\item{cmat}{A covariance or correlation matrix}
}
\description{
Stan formatted Cholesky factored cov/cor matrix
}
\examples{
x <- rnorm(10)
y <- rnorm(10)
z <- rnorm(10)

X <- cbind(x,y,z)

cmat <- cor(X)

stan_chol(cmat)
}

