% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stan_fit_stat}
\alias{stan_fit_stat}
\title{WAIC and LOO fit statistics}
\usage{
stan_fit_stat(log_lik)
}
\arguments{
\item{log_lik}{A matrix of log-likelihoods, typically from a stan model}
}
\description{
Will find the WAIC and LOO stats if given a m x n matrix of log-likelihoods, where n= n obs and m= n samples
}
\details{
You must have estimated log_lik parameter or similarly named parameter in your model
}
\examples{
stan_fit_stat(extract(rstan_pack$stan_mcmc, "log_lik")$log_lik)
}

