% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan.R
\name{stan_loo}
\alias{stan_loo}
\title{PSIS-LOO}
\usage{
stan_loo(object, par = "log_lik", plot = TRUE)
}
\arguments{
\item{object}{Stan fitted object}

\item{par}{Name of parameter that holds the log-likelihood from the model}

\item{plot}{Plot Pareto shape parameters}
}
\value{
Name list from \code{loo:loo} function
}
\description{
Pareto Smoothed Importance Sampling-Approximate Leave-One-Out Cross-Validation (PSIS-LOO)
}
\details{
See \code{?loo::loo-package} for more details on this method.
}
\examples{
stanfit <- rstan_mejr()
loo_list <- stan_loo(stanfit$stan_mcmc)
}

