% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{stan_plots_mejr}
\alias{stan_plots_mejr}
\title{RStan plots}
\usage{
stan_plots_mejr(stan_obj, pars, out, inc_warmup = FALSE)
}
\arguments{
\item{stan_obj}{The fitted object from \code{rstan::stan}}

\item{pars}{A character vector of parameter names to plot}

\item{out}{Path for where to save the plots}

\item{inc_warmup}{Include warmup for all plots. Defaults to \code{FALSE}}
}
\value{
nothing. Prints plots to disk
}
\description{
Saves a series of plots from a fitted rstan object
}
\examples{
stan_fit <- rstan_mejr()
stan_plots_mejr(stan_fit)
}

