% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{stan_pram_hist}
\alias{stan_pram_hist}
\title{Point estimates and histogram plots of fitted parameters in Stan}
\usage{
stan_pram_hist(x, fname = "plot_stanfit_hist.pdf", pars, include = TRUE,
  bndw = 1, ...)
}
\arguments{
\item{x}{rstan object}

\item{fname}{pdf file name for histograms}

\item{pars}{parameters to plot. Defaults to all}

\item{include}{if pars is used, this states whether to include or exclude those parameters.}

\item{bndw}{adjust density line bandwidth}

\item{...}{additional options passed to function \code{hdiq}}
}
\description{
Plots histograms, densities, and central tendency (defaults to median)
}
\examples{
rstan_pack <- stan_mejr()
stan_model <- rstan_pack$stan_mcmc
pram_hist(stan_model)
}

