% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stdvec}
\alias{stdvec}
\title{Get mixed-effects standard deviations}
\usage{
stdvec(model, grp = NULL)
}
\arguments{
\item{model}{Fitted model object from the \link{lme4} pacakge.}

\item{grp}{Character string naming the grouping variable used in the model formula.
Can be a vector of grouping names if more than one grouping variable.}
}
\value{
Standard deviation vector
}
\description{
Returns the standard deviation vector from a fitted model from the \code{lme4} package.
}
\details{
A model must be fitted first. If you don't specify a grouping variable name, all grouping variable standard deviatons
will be returned instead as a list. I'm not auto loading the \link{lme4} package so you have to do it yourself.
}
\examples{
library(lme4)
fm1 <- lmer(Reaction ~ Days + (Days|Subject), sleepstudy)

stdvec(fm1)
stdvec(fm1, "Subject")
}
