% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{student_t}
\alias{student_t}
\title{Student t density function}
\usage{
student_t(x, v, m = 0, s = 1, Plot = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{v}{nu (degrees of freedom) parameter}

\item{m}{mean parameter}

\item{s}{standard deviation parameter}

\item{Plot}{plot densities and don't return density vector. 
A dotted line of a normal distribution is shown for reference.}
}
\value{
a numeric vector of densities
}
\description{
Student t density function
}
\examples{
student_t(x = seq(-25,45,length.out=100), v = 2, m = 10, s = 5, Plot = TRUE)
}

