% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{students_t}
\alias{students_t}
\title{Univariate Student's-t density function}
\usage{
students_t(x, v, m = 0, s = 1, plot = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{v}{nu (degrees of freedom) parameter}

\item{m}{mean parameter}

\item{s}{standard deviation parameter}

\item{plot}{plot densities}
}
\value{
a numeric vector of densities
}
\description{
Univariate Student's-t density function
}
\examples{
# A dotted line from the normal distribution density is shown for reference.
students_t(x = seq(-25,45,length.out=100), v = 2, m = 10, s = 5, plot = TRUE)
}
