% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mejr.R
\name{test_mejr_theme}
\alias{test_mejr_theme}
\title{Test theme by printing plots to pdf and viewport}
\usage{
test_mejr_theme(w = 6.875, h = 4.5, eplot = list(),
  mejr = list(debug_text = TRUE), gg = ggplot2::theme())
}
\arguments{
\item{w}{pdf width (inches)}

\item{h}{pdf height (inches)}

\item{eplot}{list of options for \code{example_plot}}

\item{mejr}{list of options for theme_mejr(...)}

\item{gg}{further theme customization with ggplot::theme(...)}
}
\description{
Test theme by printing plots to pdf and viewport
}
\examples{
test_mejr_theme(w = 3.25, h = 3,
  mejr = list(base_size = 8, debug_text=FALSE),
  gg = ggplot2::theme(plot.title=element_blank()))
}
