% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{theme_mejr}
\alias{theme_mejr}
\title{Custom ggplot2 theme}
\usage{
theme_mejr(base_size = 11, black_level = 255, font_type = "sans",
  debug_text = FALSE)
}
\arguments{
\item{base_size}{The baseline size of text in pts. Defaults to 16.}

\item{black_level}{Values from 0 to 255, indicating the darkest line and text colors (255)}

\item{font_type}{One of the R fonts, defaults to "sans", can also use "serif"}
}
\description{
A complete, minimal theme to be used with the ggplot2 package
}
\details{
You can use \code{theme_update} to change some aspect of this theme after using \code{theme_set}.
}
\examples{
ggplot2::theme_set(theme_mejr(debug_text = TRUE))
examplePlot()

ggplot2::theme_set(theme_mejr())
ggplot2::theme_update()
plotPDF(examplePlot(), f = normalizePath(file.path("~/../Desktop/test.pdf"), mustWork = F), w = 5.5, h = 3.5)
plotPDF(examplePlot(F), f = normalizePath(file.path("~/../Desktop/test.pdf"), mustWork = F), w = 3.0, h = 2.0)
}
\seealso{
theme_update

Other graphics: \code{\link{alpha_override}},
  \code{\link{axisLim}}, \code{\link{getHCL}},
  \code{\link{getLuminance}}, \code{\link{mPlot}},
  \code{\link{marginText}}, \code{\link{rainbow2}}
}
\keyword{ggplot2}
\keyword{theme_set}

