% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general.R
\name{ts2frame}
\alias{ts2frame}
\title{Convert timestamps to frame numbers}
\usage{
ts2frame(x, fps = 30, tstart = 0, tend, chunked = FALSE, warn = TRUE)
}
\arguments{
\item{x}{Vector of timestamps}

\item{fps}{Frames per second of the video source.
Defaults to 30 Frames Per Second. The smaller the value, the more likely
two events will be chunked in the same frame.}

\item{tstart}{Start timestamp. Anything below start will not be converted.}

\item{tend}{End timestamp. Anything above will be NA. Defaults to max of x if not set.}

\item{chunked}{If set to TRUE, will return a time back to you instead of frame number,
but the chunked/cut value corresponding to that frame.}

\item{warn}{Turn on/off warnings for NAs}

\item{tstart}{time to be used as the initial level in a factor. Assumes 0 time.}
}
\description{
Provide the frame rate to use to create break points from millisecond time data
}
\examples{
# sequence of milliseconds
x <- seq(1, 1009, 12)

# 30 fps video
ts2frame(x, fps=30)

# first frames are NA until start frame is encountered
ts2frame(x, fps=29.97, tstart=333)

# compare chunked time to actual time
cbind(sprintf("\%.2f", ts2frame(x, tstart=100, tend=1000, fps=30, chunked=TRUE)), x)
}

