% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{unload_pkg}
\alias{unload_pkg}
\title{Unload package(s)}
\usage{
unload_pkg(...)
}
\arguments{
\item{...}{unquoted package names}
}
\description{
This will force unload a character vector of packages
}
\details{
You can use a vector to name more than one package to unload.
}
\examples{
library(tools)
library(mejr)
unload_pkg(c("tools", "mejr"))
}
\seealso{
Other helpers: \code{\link{clear_ws}},
  \code{\link{fac2num}}, \code{\link{has_data}},
  \code{\link{print_sec}}
}
\concept{helpers}
