% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{update_opts}
\alias{update_opts}
\title{Update R options and return old values}
\usage{
update_opts(..., opts_list = NULL)
}
\arguments{
\item{...}{key-value arguments to pass to \code{options}}

\item{opts_list}{list of options instead of using key-val args}
}
\value{
list of options with previously set values before changes
}
\description{
Update R options and return old values
}
\examples{
# key-val args have priority
update_opts(max.print=5000, width=250, opts_list=list(width=100))
getOption('width') # <- 250
}
