% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{update_opts_rmd}
\alias{update_opts_rmd}
\title{Update R Markdown options}
\usage{
update_opts_rmd()
}
\value{
list of old options
}
\description{
Update R Markdown options
}
\examples{
update_opts_rmd()
}
