% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{update_pander_opts}
\alias{update_pander_opts}
\title{update options from pander package from named list}
\usage{
update_pander_opts(opts = NULL)
}
\description{
update options from pander package from named list
}
