% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{view_data}
\alias{view_data}
\title{View a data object with ggvis}
\usage{
view_data(x, n_rows = getOption("mejr.viewdata.nrows"),
  page_size = getOption("mejr.viewdata.pagesize"),
  page_height = getOptions("mejr.viewdata.height"), subset_fun = head)
}
\description{
View a data object with ggvis
}
\examples{
view_data(iris, -1)
}
