% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stan.R
\name{view_stan_chains}
\alias{view_stan_chains}
\title{Plot each chain and return the variances}
\usage{
view_stan_chains(chainlist)
}
\arguments{
\item{chainlist}{A list of separate chains, if using \code{rstan_mejr} this is the \code{cl} list name.}
}
\description{
This is to see which chains should be removed, if any.
}
\details{
Details soon.
}
\examples{
rstan_pack <- rstan_mejr(parallel=TRUE)
view_stan_chains(rstan_pack$cl)
}

