% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{zMat}
\alias{zMat}
\title{Return a Z matrix for use with multilevel models}
\usage{
zMat(formula, x)
}
\arguments{
\item{formula}{A formula in the same form as \link{model.matrix} but using the bar syntax of lme4.}

\item{x}{the data to use the formula on.}
}
\value{
Matrix
}
\description{
If given a formula this function will return Z using that formula.
}
\examples{
library(lme4)

# formula based on lme4 sleepstudy data

form <- formula(~ 1 + Days | Subject)
z <- zMat(form, sleepstudy)
}

