% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general.R
\name{age_months}
\alias{age_months}
\title{Age calculater (months)}
\usage{
age_months(dob, ref, lub.fmt = lubridate::mdy)
}
\arguments{
\item{dob}{Date of birth string that the function \code{\link{ymd}} and others like it can understand. Typically in the format "yyyy-mm-dd" or "yyyy/mm/dd"}

\item{ref}{Reference date string. Either today's date or some other time after DOB. Defaults to today.}

\item{lub.fmt}{Lubridate function for the input dates, such as \code{\link{ymd}} or \code{\link{mdy}}}
}
\value{
Numeric value of age in months
}
\description{
Calculates ages in months with decimal days from date of birth until up to some point
}
\details{
If you're going to use a reference date, make sure the format for both dob and ref are the same. For example, don't use ymd for dob and mdy for ref. You'll get wrong values.
}
\examples{
age_months("01-10-2013")
}

