% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general.R
\name{categorize}
\alias{categorize}
\title{Categorize strings into bins}
\usage{
categorize(vec, catlist, asfactor = TRUE)
}
\arguments{
\item{vec}{The original vector that needs to be categorized.}

\item{catlist}{A list object defining the categories and levels within the category}

\item{asfactor}{Convert the final vector to a factor}
}
\description{
Take a vector of strings and categorize them according to the provided list object
}
\details{
The function works similarly to the \code{factor} function where labels are repeated.
In order to categorize numeric ranges, use the \link{cut} function instead.
Missing assignments will be marked as \code{NA}.
}
\examples{
alphabet <- letters[1:26]
classes <- list(
`first set` = letters[1:10],
`second set` = letters[15:20],
`third set` = letters[21:26]
)

categorize(alphabet, classes)
}

