% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{denseMode}
\alias{denseMode}
\title{Mode from density estimation}
\usage{
denseMode(x, adjust = 1.5, ...)
}
\arguments{
\item{x}{Value vector. Numeric or integers.}

\item{adjust}{Bandwidth adjustment. See \link{density}.}
}
\description{
Finds the mode using the \link{density} function and then obtains the maximum value.
}
\examples{
# Mixture distribution
x <- rchisq(1000, 5)
denseMode(x)
median(x)
mean(x)
}

