\name{fac2num}
\alias{fac2num}
\title{Factor to numeric class}
\usage{
fac2num(x)
}
\arguments{
  \item{x}{Factor column}
}
\value{
Numeric vector
}
\description{
Takes a factor and tries to coerce to numeric. Helpful if
numbers have been coverted to factors.
}
\seealso{
Other helpers: \code{\link{has_data}};
\code{\link{loadpkg}}; \code{\link{print_sec}}
}

