% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{font_initial_setup}
\alias{font_initial_setup}
\title{Create font database and/or load fonts}
\usage{
font_initial_setup(db_import = FALSE, gs_path = "")
}
\arguments{
\item{db_import}{create font database using extrafont}

\item{gs_path}{set ghostscript path if not using R_GSCMD}
}
\description{
Creating font database only has to be done once or if installing a new font.
Requires the \code{extrafont} package.
}
\examples{
# just load existing font database, or you can just call: library(extrafont)
font_initial_setup(FALSE)

# create database for the first time
font_initial_setup(TRUE)
}
