% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{formula2char}
\alias{formula2char}
\title{Convert formula to character}
\usage{
formula2char(x)
}
\arguments{
\item{x}{a formula}
}
\value{
string
}
\description{
Convert formula to character
}
\examples{
formula2char(y ~ x + b)
}
