% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-misc.R
\name{get_colors}
\alias{get_colors}
\title{Get colors from Brewer pallette}
\usage{
get_colors(n = 11, set = "Spectral")
}
\arguments{
\item{n}{number of colors to generate. Set to \code{NULL} to view color set names}

\item{set}{Character name of set to use, see \link{RColorBrewer::brewer.pal}}
}
\value{
A character vector of hex color codes
}
\description{
Get colors from Brewer pallette
}
\examples{
get_colors(16)
show_colors(get_colors(64), FALSE)
get_colors(NULL)
}
