% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{heat_colors}
\alias{heat_colors}
\title{A set of perceptually uniform heat map colors from python's matplotlib}
\usage{
heat_colors(n)
}
\arguments{
\item{n}{an integer for the number of colors to get}
}
\value{
Hex color codes
}
\description{
A set of perceptually uniform heat map colors from python's matplotlib
}
\examples{
heat_colors(10)
show_colors(heat_colors(36))
}
