% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{knit_lme}
\alias{knit_lme}
\title{print lme4 model}
\usage{
knit_lme(merMod, top_level = 3, digits = 3, aov = TRUE)
}
\arguments{
\item{merMod}{lmer or glmer object}

\item{top_level}{Number of #'s, defaults to 3}

\item{digits}{significant digits to print}

\item{aov}{print ANOVA, logical}
}
\description{
print lme4 model
}
\examples{
knit_lme(lme4::lmer(Reaction ~ Days + (Days | Subject), lme4::sleepstudy))
}

