% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{logmean}
\alias{logmean}
\title{Log mean}
\usage{
logmean(x)
}
\arguments{
\item{x}{numeric or integer vector, non-negative. NAs are removed automatically.}
}
\description{
Finds the geometric mean
}
\examples{
logmean(rgamma(20, .1, 1))
}

