% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{logmean}
\alias{logmean}
\title{Log mean}
\usage{
logmean(x, logValues = FALSE)
}
\arguments{
\item{x}{numeric or integer value or vector of values of these types}

\item{sum2one}{if TRUE (default) then the vector sums to 1, else each value is a proportion of the max distance.}
}
\description{
Finds the mean by first finding log or exp then back to original scale
}
\examples{
x <- runif(10, -100, 100)
normalize(x, sum2one=TRUE)
normalize(x, sum2one=FALSE)
}

