% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mejr.R
\docType{package}
\name{mejr}
\alias{mejr}
\alias{mejr-package}
\title{mejr: Package}
\description{
A package to distribute a set of functions and operations that I've
found to be useful over time.
}
\section{Info}{

See \code{library(help="mejr")} for a list of functions. \cr
\cr
Some useful links and external documentation \cr
\cr
\url{http://docs.ggplot2.org/current/}
\cr
}

