% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mejr.R
\docType{package}
\name{mejr}
\alias{mejr}
\alias{mejr-package}
\title{mejr: Package}
\description{
A package to distribute a set of functions and perform common operations that
I've found to be useful or time saving.
}
\section{Info}{
 See \code{library(help="mejr")} for a list of functions. \cr
  \cr Some useful links and external documentation \cr \cr
  \url{http://ggplot2.tidyverse.org/reference/} \cr
}

