% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mejr.R
\docType{package}
\name{mejr}
\alias{mejr}
\alias{mejr-package}
\title{Mejr Package}
\description{
This package has no main purpose but to make my life easier
instead of copying and pasting functions from different folder.
}
\details{
See \code{library(help="mejr")} for a list of functions.

The package depends on \code{ggplot2}.

Below are some useful links I can click on instead of google searching them:

\url{http://docs.ggplot2.org/current/}
}

