% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ms2frames}
\alias{ms2frames}
\title{Convert milliseconds to frame number}
\usage{
ms2frames(x, fps = 30, tstart = 0)
}
\arguments{
\item{x}{millisecond numeric or integer data}

\item{fps}{frames per second. Single value.}

\item{tstart}{time to be used as the initial level in a factor. Assumes 0 time.}
}
\description{
Provide the frame rate to use to create break points from millisecond time data
}
\details{
Assumes time starts at 0, but this can be changed.
}
\examples{
ms2frames(x=c(999, 1333), fps=30)
ms2frames(x=c(999, 1333), fps=30, tstart=333)
}

