% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general.R
\name{normalize}
\alias{normalize}
\title{Normalize a vector of values}
\usage{
normalize(x, sum2one = TRUE)
}
\arguments{
\item{x}{numeric or integer value or vector of values of these types}

\item{sum2one}{if TRUE (default) then the vector sums to 1, else each value is a proportion of the max distance.}
}
\description{
Normalize a vector of values
}
\details{
Will normalize to sum to one or take the max and use that as one.
}
\examples{
x <- runif(10, -100, 100)
normalize(x, sum2one=TRUE)
normalize(x, sum2one=FALSE)
}

