% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{normalize}
\alias{normalize}
\title{Normalize a vector of values}
\usage{
normalize(x, type = c("one", "zero", "max"))
}
\arguments{
\item{x}{numeric or integer value or vector of values of these types}

\item{type}{character of the type of normalization to perform. Defaults to "one"}
}
\description{
Normalize to sum to one, sum to zero, or as a proportion of max value
}
\examples{
x <- sort(runif(10, -100, 100))
normalize(x, "one")  # all values will sum to one
normalize(x, "zero") # all values will sum to zero
normalize(x, "max")  # all values divided by most extreme value
}

