% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{onion_chol}
\alias{onion_chol}
\title{Covariance matrix using onion method}
\usage{
onion_chol(tau, zeta, rho, z, scale = NULL, dispersion = NULL,
  as.vec = FALSE)
}
\arguments{
\item{tau}{trace (sum of variances)}

\item{zeta}{conjugate to dirichlet distribution}

\item{rho}{correlation sign and magnitude}

\item{z}{correlation scaler (-Inf, Inf)}

\item{scale}{scale variances}

\item{dispersion}{dispersion of errors}

\item{as.vec}{return vector}
}
\value{
matrix
}
\description{
Covariance matrix using onion method
}
\examples{
p <- 4
chol_cov <- onion_chol(
  tau = rgamma(1, shape = 1.1, scale = 1), 
  zeta = rgamma(p, shape = 2, scale = 1), 
  rho = rbeta(p - 1, 1, 0.8),
  z = rnorm(choose(p, 2) - 1), 
  scale = 5)

tcrossprod(chol_cov)
}
