% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan.R
\name{pram_hist}
\alias{pram_hist}
\title{Point estimates and histogram plots of fitted parameters in Stan}
\usage{
pram_hist(x, bndw = 0.33, fname = "plot_stanfit_hist.pdf",
  print_hist = TRUE)
}
\arguments{
\item{x}{rstan object}

\item{bndw}{adjust density line bandwidth}

\item{fname}{pdf file name for histograms}
}
\description{
Plots histograms, densities, and central tendency (defaults to median)
}
\examples{
rstan_pack <- rstan_mejr(parallel=FALSE)
stan_model <- rstan_pack$stan_mcmc
pram_hist(stan_model)
}

