% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{print_sec}
\alias{print_sec}
\title{Prints a section title to console}
\usage{
print_sec(x, docwidth = 80)
}
\arguments{
\item{x}{Section title}

\item{docwidth}{Character width (line width)}
}
\value{
NULL, prints to console or sink
}
\description{
When printing contents to a file, use this to mark sections of code
}
\examples{
print_sec()

print_sec("Results")
}
\seealso{
Other helpers: \code{\link{clear_ws}},
  \code{\link{fac2num}}, \code{\link{has_data}},
  \code{\link{unload_pkg}}
}
\keyword{section}
