% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general.R
\name{printSec}
\alias{printSec}
\title{Prints a section title to console}
\usage{
printSec(x, char = 80)
}
\arguments{
\item{x}{Section title}

\item{char}{Character width (line width)}
}
\value{
NULL, prints to console or sink
}
\description{
When printing contents to a file, use this to mark sections of code
}
\examples{
printSec()

printSec("Results")
}
\seealso{
Other helpers: \code{\link{clrAll}}; \code{\link{fac2num}};
  \code{\link{hasData}}; \code{\link{loadPkg}};
  \code{\link{rmPkg}}
}
\keyword{section}

