% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rcov}
\alias{rcov}
\title{Random covariance matrix}
\usage{
rcov(n, size, regularization = 1, concentration = 1, shape = 1,
  scale = 1, scale_mat = 1/sqrt(2))
}
\arguments{
\item{n}{Number of random matrices to generate}

\item{size}{Number of columns or variances in a covariance matrix.}

\item{regularization}{Positive scalar. Controls correlation strength. 1 is
uniform over correlation matrices. Greater than 1 has weaker correlations.}

\item{concentration}{Positive scalar. Controls proportions of total
variance. 1 is uniform, less than 1 creates heterogeneity, greater than 1
makes variances homogeneous.}

\item{shape}{Positive scalar. Controls total variance. Shape of a gamma
distribution, defaults to exponential. Dispersion around mean variance}

\item{scale}{Positive scalar. Controls total variance. Scale of a gamma
distribution, defaults to exponential. Mean variance}

\item{scale_mat}{Positive scalar. Scale entire matrix by this amount.}
}
\value{
A matrix or vector if return_vector is set to TRUE
}
\description{
Random covariance matrix
}
\examples{
rcov(3, 4)
}
