% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rmPkg}
\alias{rmPkg}
\title{Unload package}
\usage{
rmPkg(pkgs)
}
\arguments{
\item{pkgs}{A character string or vector of package names}
}
\value{
NA
}
\description{
This will force unload a vector of packages
}
\details{
You can use a vector to name more than one package to unload.
}
\examples{
library(tools)
library(mejr)
rmPkg(c("tools", "mejr"))
}
\seealso{
Other helpers: \code{\link{clrAll}}; \code{\link{fac2num}};
  \code{\link{hasData}}; \code{\link{loadPkg}};
  \code{\link{printSec}}
}

