% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan.R
\name{rstan_mejr}
\alias{rstan_mejr}
\title{Fit STAN model using parallel chains}
\usage{
rstan_mejr(modDat, modFile, modOpts, modPrams, modInits, modCtrl, out,
  parallel = FALSE, debug = FALSE, repackage = NULL)
}
\arguments{
\item{modDat}{model data as a list}

\item{modFile}{character vector of a path that points to the model file to use}

\item{modOpts}{list of options. If none are provide defaults to:
\code{list(n_chains = 3, n_final = 2000, n_thin = 2, n_warm = 500)}}

\item{modPrams}{character vector of parameters to keep. Defaults to all.}

\item{modInits}{initializations function to use}

\item{modCtrl}{control parameters from stan function}

\item{out}{character vector specificy path to output all files}

\item{parallel}{whether to run parallel chains using package \code{snow}. Must install this package first.}

\item{debug}{write arguments to .GlobalEnv for debugging}

\item{repackage}{add on R object to the final list object. This will be added to \code{stan_fit.RData} as \code{rstan_pack$repack}}
}
\description{
This is a convenience wrapper for the \code{stan} function to allow for parallel chains and other default options.
}
\details{
Details soon.
}
\examples{
# Run default example model
rstan_pack <- rstan_mejr(parallel=FALSE)
}

