% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{save_plot}
\alias{save_plot}
\title{Custom plotting function to save PDF and PNG files}
\usage{
save_plot(x, file, dir = NULL, width = 5.25, height = 3.8,
  format = "pdf", font = getOption("mejr.font"), onefile = FALSE,
  res = 300, fun = NULL, ...)
}
\arguments{
\item{x}{a plot object or list of plots}

\item{file}{file name for plot. Uses name of object p as default.}

\item{dir}{directory of where to save plot. Defaults to current working directory.}

\item{width}{width of plot in inches}

\item{height}{height of plot in inches}

\item{format}{can be "pdf", "png", or "both"}

\item{font}{name of font family to embed into file}

\item{onefile}{print to a single pdf device or multiple}

\item{res}{png pixel resolution}

\item{fun}{function to use before dev.off() is called. Can print other stuff to output.}

\item{...}{args passed to fun}
}
\description{
Creates plots using some default settings
}
\details{
Automatically open and closes the graphics device after use.
}
\examples{
my_plots <- list(hist(rnorm(100)), hist(rpois(100, 10)))
save_plot(my_plots, dir = "~/../Desktop", format = "both")

# embed font (extrafont package)
library(extrafont)
library(ggplot2)
custom_font_plot <- example_plot()+theme_mejr(font_family='Franklin Gothic Medium')
save_plot(custom_font_plot, dir = "~/../Desktop", format = "pdf", font='Franklin Gothic Medium')
}
