% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sd2}
\alias{sd2}
\title{Custom standard deviation}
\usage{
sd2(x, fun = mean, correction = 1.5, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{fun}{summary statistic, measure of central tendency. Defaults to \code{mean}}

\item{correction}{bias correction amount, defaults to 1.5 instead of 1}

\item{...}{additional arguments passed to fun}
}
\value{
scalar
}
\description{
Inject a different summary statistic for the mean and adjust the bias correction term
}
\examples{
x <- rpois(100, 25)
sd(x)                    # standard
sd2(x, correction = 1)   # same as above
sd2(x)                   # 1.5 bias correction
sd2(x, logmean, 1)       # geometric mean, correction=1
sd2(x, logmean)          # geometric mean, correction=1.5
}

