% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{show_colors}
\alias{show_colors}
\title{plot and show hex values of colors}
\usage{
show_colors(colors, show.legend = TRUE, cols = NULL)
}
\arguments{
\item{colors}{character vector of hex value colors}

\item{show.legend}{show the legend with hex values (logical)}
}
\value{
A plot with the index of the color in the tile
}
\description{
plot and show hex values of colors
}
\examples{
show_colors(color_10(5))
show_colors(get_colors(25))
show_colors(get_colors(64), FALSE)
}
