% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stats.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoidal (logistic) function}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{Numeric value or vector of values on the logistic scale.}
}
\value{
Numeric values ranging from 0 to 1
}
\description{
One-liner of the sigmoidal (logistic) function.
Typically used as a link for the linear predictor of a glm.
}
\examples{
sigmoid(0)
sigmoid(10)

x <- seq(-10,10,.5)
plot(x=x, y=sigmoid(x), type="l")
}

