% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid function}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{Numeric value or vector of values on the logit (log-odds) scale.}
}
\value{
Numeric values ranging from 0 to 1
}
\description{
Compress logit scaled values between 0 and 1.
}
\examples{
sigmoid(-4.595)
sigmoid(11.5)

x <- logit(seq(0.001, 0.999, length.out = 50))
plot(x = x, y = sigmoid(x), type = "l")
}
