% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{symbol2char}
\alias{symbol2char}
\title{Convert symbol/equation to character}
\usage{
symbol2char(...)
}
\arguments{
\item{...}{expression}
}
\value{
list of character strings
}
\description{
Convert symbol/equation to character
}
\examples{
symbol2char(y ~ x + z, y ~ x + x^2, (. ~ .), "'\\" \\"'")
}
\seealso{
dots2list
}
