% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{theme_mejr}
\alias{theme_mejr}
\title{Custom ggplot2 theme}
\usage{
theme_mejr(base_size = 9, black_level = 204, font_type = "sans",
  debug_text = FALSE)
}
\arguments{
\item{base_size}{The baseline size of text in pts. Defaults to 16.}

\item{black_level}{Values from 0 to 255, indicating the darkest line and text colors (255)}

\item{font_type}{One of the R fonts, defaults to "sans", can also use "serif"}
}
\description{
A complete, minimal theme to be used with the ggplot2 package
}
\details{
You can use \code{theme_update} to change some aspect of this theme after using \code{theme_set}.
}
\examples{
ggplot2::theme_set(theme_mejr(debug_text = TRUE))
example_plot()

ggplot2::theme_set(theme_mejr())
ggplot2::theme_update()          # any updates can go here
save_plot(example_plot(), file = normalizePath(file.path("~/../Desktop/test"), mustWork = F))
save_plot(example_plot(F),
 file = normalizePath(file.path("~/../Desktop/test"), mustWork = F),
 width = 3.0, height = 2.0)
}
\seealso{
theme_update

Other graphics: \code{\link{alpha_override}},
  \code{\link{axis_limit}}, \code{\link{getHCL}},
  \code{\link{getLuminance}}, \code{\link{margin_text}},
  \code{\link{rainbow_colors}}
}
\keyword{ggplot2}
\keyword{theme_set}
