% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mejr.R
\name{theme_mejr}
\alias{theme_mejr}
\title{Custom ggplot2 theme}
\usage{
theme_mejr(base_size = 11, madj = NULL, black_level = 204,
  font_family = "sans", debug_text = FALSE)
}
\arguments{
\item{base_size}{The baseline size of text in pts. Defaults to 16.}

\item{madj}{additive adjustment of margin spacing}

\item{black_level}{Values from 0 to 255, indicating the darkest line and text colors (255)}

\item{font_family}{One of the R fonts, defaults to "sans", can also use "serif"}

\item{debug_text}{show text placement with yellow highlight}
}
\description{
A complete, minimal theme to be used with the ggplot2 package
}
\details{
You can use \code{theme_update} to change some aspect of this theme after using \code{theme_set}.
}
\examples{
ggplot2::theme_set(theme_mejr(debug_text = TRUE))
example_plot()

ggplot2::theme_set(theme_mejr())
ggplot2::theme_update(plot.title = element_blank()) # any updates can go here
example_plot()
}
\seealso{
theme_update

Other graphics: \code{\link{alpha_override}},
  \code{\link{luminance}}
}
\keyword{ggplot2}
\keyword{theme_set}
