% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trim}
\alias{trim}
\title{Trim extreme values}
\usage{
trim(x, tr = 0.05, rm.na = TRUE)
}
\arguments{
\item{x}{Vector of numeric values.}

\item{tr}{How much to trim as a proportion}

\item{rm.na}{Set to FALSE to keep NA values in the output vector}
}
\description{
This will trim out extreme values and return the same order as the input vector with values removed.
}
\examples{
x <- rpois(1000, 15)
trim(x, tr=0.1)
}

