% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{view_data}
\alias{view_data}
\title{View a data object with ggvis}
\usage{
view_data(x, n_rows = getOption("mejr.viewdata.nrows"),
  page_size = getOption("mejr.viewdata.pagesize"),
  page_height = getOption("mejr.viewdata.height"),
  subset_fun = getOption("mejr.viewdata.subsetfun"))
}
\description{
View a data object with ggvis
}
\examples{
# all data
view_data(iris, 0)

# all data minus last 2
view_data(iris, -2)

# these rows
view_data(iris, 10:15)
}
