% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan.R
\name{view_stan_chains}
\alias{view_stan_chains}
\title{Plot each chain and return convergence stats}
\usage{
view_stan_chains(stanmodel, pars)
}
\arguments{
\item{stanmodel}{A fitted stan model}

\item{pars}{Names of parameters to calculate convergence statistics, defaults to all parameters, including lp__}
}
\description{
This is to see which chains should be removed, if any.
}
\details{
If you get an error and using RStudio, try to make plot window bigger.
}
\examples{
rstan_pack <- rstan_mejr(parallel=TRUE)
view_stan_chains(rstan_pack$stan_mcmc)
}

